+++ packages/backend-function/src/function_env_translator.ts
@@ -4,6 +4,7 @@ import { FunctionProps } from './factory.js';
 import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 import { Effect, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 import { FunctionEnvironmentTypeGenerator } from './function_env_type_generator.js';
+import { AmplifyUserError } from '@aws-amplify/platform-core';
 
 /**
  * Translates function environment props into appropriate environment records and builds a policy statement
@@ -35,7 +36,14 @@ export class FunctionEnvironmentTranslator {
           `${this.amplifySsmEnvConfigKey} is a reserved environment variable name`
         );
       }
-      if (typeof value !== 'string') {
+      if (typeof value === 'undefined') {
+        throw new AmplifyUserError('InvalidFunctionConfiguration', {
+          message: `The value of environment variable ${key} is undefined.`,
+          resolution: `Ensure that all defineFunction environment variables are defined.`,
+        });
+      } else if (typeof value === 'string') {
+        this.lambda.addEnvironment(key, value);
+      } else {
         const { branchSecretPath, sharedSecretPath } =
           this.backendSecretResolver.resolvePath(value);
         this.lambda.addEnvironment(key, this.ssmValuePlaceholderText);
@@ -44,8 +52,6 @@ export class FunctionEnvironmentTranslator {
           sharedPath: sharedSecretPath,
         };
         this.ssmPaths.push(branchSecretPath, sharedSecretPath);
-      } else {
-        this.lambda.addEnvironment(key, value);
       }
       this.amplifyBackendEnvVarNames.push(key);
     }